+++ src/lib_reloader.rs
@@ -15,6 +15,10 @@ use std::thread;
 use std::time::Duration;
 
 use crate::error::HotReloaderError;
+use crate::util::TimedLimiter;
+
+#[cfg(feature = "verbose")]
+use log;
 
 /// Manages watches a library (dylib) file, loads it using
 /// [`libloading::Library`] and [provides access to its
@@ -38,6 +42,8 @@ pub struct LibReloader {
     loaded_lib_file: PathBuf,
     lib_file_hash: Arc<AtomicU32>,
     file_change_subscribers: Arc<Mutex<Vec<mpsc::Sender<()>>>>,
+    #[cfg(target_os = "macos")]
+    codesigner: crate::codesign::CodeSigner,
 }
 
 impl LibReloader {
@@ -60,6 +66,9 @@ impl LibReloader {
 
         let load_counter = 0;
 
+        #[cfg(target_os = "macos")]
+        let codesigner = crate::codesign::CodeSigner::new();
+
         let (watched_lib_file, loaded_lib_file) =
             watched_and_loaded_library_paths(&lib_dir, &lib_name, load_counter);
 
@@ -68,6 +77,8 @@ impl LibReloader {
             // where a file lock would be held, preventing the lib from changing later.
             log::debug!("copying {watched_lib_file:?} -> {loaded_lib_file:?}");
             fs::copy(&watched_lib_file, &loaded_lib_file)?;
+            #[cfg(target_os = "macos")]
+            codesigner.codesign(&loaded_lib_file);
             (
                 hash_file(&loaded_lib_file),
                 Some(load_library(&loaded_lib_file)?),
@@ -98,6 +109,8 @@ impl LibReloader {
             lib_file_hash,
             changed,
             file_change_subscribers,
+            #[cfg(target_os = "macos")]
+            codesigner,
         };
 
         Ok(lib_loader)
@@ -154,6 +167,8 @@ impl LibReloader {
                 watched_and_loaded_library_paths(lib_dir, lib_name, *load_counter);
             log::trace!("copy {watched_lib_file:?} -> {loaded_lib_file:?}");
             fs::copy(watched_lib_file, &loaded_lib_file)?;
+            #[cfg(target_os = "macos")]
+            self.codesigner.codesign(&loaded_lib_file);
             self.lib_file_hash
                 .store(hash_file(&loaded_lib_file), Ordering::Release);
             self.lib = Some(load_library(&loaded_lib_file)?);
@@ -188,8 +203,19 @@ impl LibReloader {
                 .watch(&lib_file, RecursiveMode::NonRecursive)
                 .expect("watch lib file");
 
-            let signal_change = || {
-                if hash_file(&lib_file) == lib_file_hash.load(Ordering::Acquire) {
+            // Some changes such as codesigning can emit subsequent file change
+            // events that can be received delayed. Those wont get caught by the
+            // `debounce` delay of the watcher as we ourselve are responsible
+            // for those changes happening. As we don't want to run into a reload
+            // loop we limit the amount of signaled changes that can happen in a
+            // short timespan additionally with the `TimedLimiter`.
+            let mut signal_limit = TimedLimiter::<3>::new(Duration::from_secs(1));
+
+            let mut signal_change = || {
+                if signal_limit.too_frequent()
+                    || hash_file(&lib_file) == lib_file_hash.load(Ordering::Acquire)
+                    || changed.load(Ordering::Acquire)
+                {
                     // file not changed
                     return false;
                 }